Fix+tests for newick with whitespace
Added test for parsing a newick string with whitespace, then fixed the bug that showed
up because of said test.
Fix: the float value from the pest parser needed to be trimmed to be a viable float,
otherwise the branch length was being set to 0 due to the whitespace.
